move interface to separate file
